Add components necessary for SMPS using A8498
The A8498 can supply up to 3A, and the IC has a current limit of 3.5-5A.
Although it's a bit vague it is high enough to power 6 servos under full load.
The current of an average servo is about 600mA when fully loaded according to
some 'Hobby Servo Fundmentals' PDF I found on the internet. The likely hood of
all 8 servos being fully loaded at the same time is very low and a higher
current SMPS would increase the cost and complexity of the circuit required
significantly, 3A seems to be some kind of limit for ICs with integrated FETs.