Add power LED to servo 5V rail. Add zener clamped sense inputs to MSP430
Both the servo 5V rail and servo ground rail are connected to the MSP430,
clamped via zener diodes, to allow the detection of power and also dodgy
wiring. If the power is connected the wrong way around then both the 'S5V' and
'SGND' nodes will be >3.3V so both inputs will be detectable as high. Under
normal circumstances only the 'S5V' input should be high. The 'SGND' input
indicates a fault.