Add some 1k resistors between the MSP and the buffers
I think it could be possible for 12V to be injected into the MSP if
12V is connected to both of the power camcon terminals. All of the
stuff on the high current side would be brought up to 12V, including
the ground net. Current could then flow from the 'SGND' net, through
the buffer input clamping diodes, into the MSP. The 1k resistors will
limit this current.