Change the inverter/buffers between the MSP and servos to a 3-8 decoder
There was always problems writing the firmware with the old design as it
requires code to run at very specific times which is not possible when there's
interrupts being fired as data comes in over I2C/SRIC. This led to the I2C/SRIC
code having to be poll based, with the only interrupt being the timer for
controlling the servos, without this they would jitter and be useless.

This design uses a 3-8 decoder to direct the output of Timer B to an individual
servo. The timer can control the width of the pulse sent to each servo very
precisely. Each servo is selected in turn and a pulse sent. This process is
repeated continuously.