Use the correct ground net to pull inputs to the 3-8 decoder low
The 'servo' side of things has its own ground net 'SGND' and this
should be used to pull inputs to the 3-8 decoder low, not the normal
'GND' net.

Both of these nets will essentially be at the same potential under
normal operation but the SGND net might be at 12V if the power connector
is miswired.