Move and change value of cap in feedback network of SMPS
After building a prototype I found that the SMPS was acting strangely.
Rather than turning the FET on solidly at the start of a cycle it would
oscillate at around 5MHz until the output has risen sufficiently. A large
amount of this 5MHz signal was visible on both the output and the feedback
line going into the control IC. I suspect this was being caused by the inductor
inducing some form of noise into the feedback line from its magnetic field,
the feedback network is positioned directly underneath the inductor. Moving the
cap in the feedback network to the lower half forms a low-pass filter and
stops the 5MHz signal completely.

I've had this configuration running for over an hour outputting more than 3A.
I've also had it outputting around 5.2A for a good few minutes and it's been
fine with it.
also seen this circuit