Add panels to support the PCB around the servo headers
As there are only connectors on one edge of the PCB and they're quite far away
from the servo headers there will be nothing supporting the PCB. The added
panels provide support to the PCB both from above and below. This should stop
it moving about both during plugging and unplugging servos.